requirements.txt
# Core dependencies
pydantic>=2.0.0
jsonschema>=4.0.0
requests>=2.31.0
pyyaml>=6.0.0

# Development dependencies (install with: pip install -r requirements.txt -r requirements-dev.txt)
# See requirements-dev.txt for development dependencies



